Generic multi-endpoint support for upstreaming
This change makes multi-endpoint support more generic, with
the motivation of making this feature useful for upstream.

The LUA script 'multiple-endpoints.lua' allows for specifying
an arbitrary number of HTTP(S) endpoints to include in the
benchmark. Endpoints will be connected to in a random, evenly
distributed fashion. After a run finished, the overall latency
will be reported (i.e. there's currently no break-down per
endpoint).

The main purpose of running a benchmark over multiple endpoints
is to allow benchmarking of e.g. a whole web application instead
of the pages and/or restful resources that make up said
application individually.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>